Add a gitignore file
You can use this file to ignore files you don't want in your git repo.
That's basically anything the average user would be surprised to see
floating about. In this case I'm surprised to see the code-workspace
file in here, since it's specific to your setup and not something
essential for this project.