Remove database from git and ignore
I'm also removing your database from git here. When you get ready
to put this application out on the internet you're definitely not
going to want to use sqlite. I'd recommend postgresql.

Aside from that you wouldn't generally commit the whole database
along side your application, as:
- It changes frequently
- It's quite large (or will be)
- It presents a security risk

When I originally made this commit I forgot to add the gitignore
file to it, so I staged the gitignore updates and then used this
command to include it in the previous commit:
`git commit --amend`