Validate title an email at the database level
When you really want to make sure a value is in the new record before
you save it you usually want to check that it's not null at a database
level. You still want to use the software validations, but there are
ways to bypass those validations inadvertently.

If you use `update_attributes` instead of `update` for example your
model validations will be skipped.

This way you can be certain that your database itself will actually
throw an error when something is going to sneak in there.